requirements.txt
# Person A: Braille Recognition Pipeline - Dependencies

# Core Data Science
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data I/O
h5py>=3.0.0
pyyaml>=5.4.0

# Evaluation
jiwer>=2.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
opencv-python>=4.8.0

# Development
jupyter>=1.0.0
ipython>=7.0.0

# Optional but recommended
scikit-learn>=0.24.0
tqdm>=4.60.0
